requirements.txt
fastapi
uvicorn[standard]
python-multipart
websockets
python-dotenv
twilio
deepgram-sdk
